+++ website/src/pages/guides/continuous-developer-onboarding/index.tsx
@@ -10,7 +10,6 @@ const ContinuousDevOnboarding: FunctionComponent<PageProps> = props => {
         portalId: '2762526',
         formId: '35e18409-5be7-4fcb-aa57-8152b34eef66',
         targetId: 'form',
-        chiliPiper: false,
         onFormSubmitted: () => window.open('/guides/dev-onboarding/sg-continuous-developer-onboarding.pdf'),
     })
