+++ website/src/pages/resources/universal-code-search-ebook.tsx
@@ -10,7 +10,6 @@ export const UniversalCodeSearchEbook: FunctionComponent<PageProps> = props => {
         portalId: '2762526',
         formId: 'd2425310-4f8c-4b6c-8b63-c4729912df82',
         targetId: 'hubspotContactForm',
-        chiliPiper: false,
     })
 
     return (
